Interactive 3-D association landscape for many phenotypes

Interactive 3-D association landscape for many phenotypes

landscapefast(
  d,
  sliceval = 7,
  chromosome = FALSE,
  pop = "GBR",
  R2 = 0.75,
  D = 0.75,
  calculateLD = FALSE,
  mutualLD = FALSE,
  phenos,
  genemap = FALSE,
  geneview = FALSE,
  levelsdown = 0
)

Arguments

d

DataFrame output from processphegwas

sliceval

Integer to indicate value of -log10(p) to do the sectionalcut. Usually value > -log10 6 is considered to be significant

chromosome

Integer to indicate the chromosome number thats interested, If not given entire chromosome is given

pop

The population to select for calculation the LD (default GB)

R2

The value to set to calculate LD

D

THe value to set to calcualte LD

calculateLD

This shoudld be set to true if the calcualte LD logic needed to be added to the plot

mutualLD

Calcualte the mutual LD SNP between the phenotypes

phenos

Vector of names of dataframes that need to do PheGWAS on. Arrange the dataframe in the order how the the phenotypes should align in y axis chromosome view the max peak is selected

genemap

if true it map SNP's to gene. This takes time as its using BioMart package to map genes.

geneview

This checks for the common genes across the section

levelsdown

Used to find the independant signals

Author

George Gittu

Examples

if (FALSE) {
# here y is output from function fastprocessphegwas
# 3D landscape-viz of all the phenos across the base pair positions(threshold > -log10(p) 7.5)
# Without iPheGWAS you get output in the order you pass phenos
landscapefast(y,sliceval = 10,phenos =phenos)

# 3D landscape visualization after applying iPheGWAS
# Passing the order from the iPheGWAS function
landscapefast(y,sliceval = 10,phenos = iphegwas(phenos))

#3D landscape visualization of chromosome number 19 (above a threshold of -log10 (p) 7.5)
landscapefast(y,sliceval = 7.5,chromosome = 19,phenos =phenos)

# 3D landscape-viz of chromosome 19, gene view active(threshold > -log10(p) 7.5)
landscape(y,sliceval = 7.5,chromosome = 19, geneview = TRUE,phenos =phenos)

# 3D landscape-viz with calculate-LD and mutual-LD functionality active
landscapefast(y, sliceval = 30, chromosome = 19,calculateLD= TRUE,mutualLD = TRUE,phenos =phenos)
}